+++ data/tasks.txt
@@ -8,4 +8,3 @@ D | 0 | return book | 2/12/2019 1800
 E | 0 | task | 12/12/2023 1400 | 12/12/2023 1600
 T | 0 | hi
 T | 0 | hi
-T | 0 | hi
