+++ src/main/java/bing/command/ByeCommand.java
@@ -1,3 +1,9 @@
+package bing.command;
+
+import bing.storage.Storage;
+import bing.task.TaskList;
+import bing.ui.Ui;
+
 public class ByeCommand implements Command {
 
     @Override
