+++ src/main/java/bing/command/DeadlineCommand.java
@@ -1,3 +1,7 @@
+package bing.command;
+
+import bing.task.Deadline;
+
 import java.time.LocalDateTime;
 
 public class DeadlineCommand extends AddCommand {
