+++ src/main/java/bing/command/EventCommand.java
@@ -1,3 +1,7 @@
+package bing.command;
+
+import bing.task.Event;
+
 import java.time.LocalDateTime;
 
 public class EventCommand extends AddCommand {
