+++ src/main/java/bing/command/MarkCommand.java
@@ -1,3 +1,11 @@
+package bing.command;
+
+import bing.storage.Storage;
+import bing.task.Task;
+import bing.task.TaskList;
+import bing.task.TaskStatus;
+import bing.ui.Ui;
+
 import java.io.IOException;
 public class MarkCommand implements Command {
     private int index;
