+++ src/main/java/bing/command/UnmarkCommand.java
@@ -1,15 +1,23 @@
+package bing.command;
+
+import bing.storage.Storage;
+import bing.task.Task;
+import bing.task.TaskList;
+import bing.task.TaskStatus;
+import bing.ui.Ui;
+
 import java.io.IOException;
-public class DeleteCommand implements Command {
+public class UnmarkCommand implements Command {
     private int index;
 
-    public DeleteCommand(int index) {
+    public UnmarkCommand(int index) {
         this.index = index;
     }
 
     @Override
     public void execute(TaskList tasks, Ui ui, Storage storage) {
         Task task = tasks.get(index);
-        tasks.deleteTask(index);
+        task.setStatus(TaskStatus.UNDONE);
         ui.showTasks(tasks);
         try {
             storage.save(tasks.getTasks());
