+++ src/main/java/bing/task/Event.java
@@ -1,3 +1,8 @@
+package bing.task;
+
+import bing.task.Task;
+import bing.task.TaskStatus;
+
 import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
