+++ src/main/java/bing/task/Task.java
@@ -1,3 +1,5 @@
+package bing.task;
+
 import java.time.format.DateTimeFormatter;
 public abstract class Task {
     private String info;
