+++ src/main/java/bing/task/TaskList.java
@@ -1,3 +1,5 @@
+package bing.task;
+
 import java.util.ArrayList;
 
 public class TaskList {
