+++ text-ui-test/EXPECTED.TXT
@@ -1,5 +1,5 @@
 ______________________________
-Hi! My name is Bing
+Hi! My name is bing.Bing
 How can I help you?
 ______________________________
